# Change logging formatting to lazily evaluate (performance critical)