@OneRaynyDay 2020-12-22
> @OneRaynyDay just making sure -- have you filed the CLA already?

Yep, I sent an email to cla@prestosql.io regarding this PR with a signed document for CLA. Let me know if you need anything on my side before we merge it :)